+++ src/command/types.rs
@@ -545,6 +545,12 @@ pub(crate) enum Command {
     Clear,
     Save,
     AofRewrite,
+    Multi,
+    Exec,
+    Discard,
+    Transaction {
+        commands: Vec<Command>,
+    },
     Info,
     Help,
     Exit,
@@ -671,6 +677,9 @@ impl Command {
             Self::Clear => "CLEAR",
             Self::Save => "SAVE",
             Self::AofRewrite => "AOFREWRITE",
+            Self::Multi => "MULTI",
+            Self::Exec | Self::Transaction { .. } => "EXEC",
+            Self::Discard => "DISCARD",
             Self::Info => "INFO",
             Self::Help => "HELP",
             Self::Exit => "EXIT",
@@ -1038,6 +1047,10 @@ impl Command {
             | Self::Len
             | Self::Save
             | Self::AofRewrite
+            | Self::Multi
+            | Self::Exec
+            | Self::Discard
+            | Self::Transaction { .. }
             | Self::Info
             | Self::Help
             | Self::Exit => return None,